server/requirements.txt
python-dotenv==1.0.0
flask==2.3.3
flask-cors==4.0.0
numpy==1.24.4
librosa==0.10.1
scipy==1.11.2